+++ source/main/ipc/index.js
@@ -9,6 +9,7 @@ import getGpuStatus from './get-gpu-status';
 import { handleBugReportRequests } from './bugReportRequestChannel';
 import { handleFileMetaRequests } from './generateFileMetaChannel';
 import { handlePaperWalletRequests } from './generatePaperWalletChannel';
+import { handleAddressPDFRequests } from './generateAddressPDFChannel';
 import { handleRewardsCsvRequests } from './generateRewardsCsvChannel';
 import { openExternalUrlChannel } from './open-external-url';
 import { openLocalDirectoryChannel } from './open-local-directory';
@@ -23,6 +24,7 @@ export default (window: BrowserWindow) => {
   handleBugReportRequests();
   handleFileMetaRequests();
   handlePaperWalletRequests();
+  handleAddressPDFRequests();
   handleRewardsCsvRequests();
   // eslint-disable-next-line no-unused-expressions
   openExternalUrlChannel;