+++ storybook/stories/wallets/receive/WalletReceive.stories.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { text, boolean, number } from '@storybook/addon-knobs';
+import { number, boolean } from '@storybook/addon-knobs';
 
 // Assets and helpers
 import WalletsWrapper from '../_utils/WalletsWrapper';
@@ -10,31 +10,40 @@ import { generateAddress } from '../../_support/utils';
 
 // Screens
 import WalletReceive from '../../../../source/renderer/app/components/wallet/receive/WalletReceive';
+import WalletReceiveDialog from '../../../../source/renderer/app/components/wallet/receive/WalletReceiveDialog';
 import VerticalFlexContainer from '../../../../source/renderer/app/components/layout/VerticalFlexContainer';
 
 storiesOf('Wallets|Receive', module)
   .addDecorator(WalletsWrapper)
-  .add('Receive', () => (
-    <VerticalFlexContainer>
-      <WalletReceive
-        walletAddress={text(
-          'Wallet address',
-          'DdzFFzCqrhsg9ngNRhHEa49se7qEMKyubT9tcE13Fkvh8QC82trpTDsNvdQV7mg9SCZiuENkf77zrtwPXrTyGMNznUsSinPC1gb2ZCqK'
+  .add('Receive', ({ currentLocale }: { currentLocale: string }) => {
+    const isIncentivizedTestnet = boolean('isIncentivizedTestnet', false);
+    const showDialog = boolean('showDialog', false);
+    return (
+      <VerticalFlexContainer>
+        <WalletReceive
+          walletAddresses={[
+            ...Array.from(Array(number('Addresses', 10))).map(() =>
+              generateAddress()
+            ),
+            ...Array.from(Array(number('Addresses (used)', 10))).map(() =>
+              generateAddress(true)
+            ),
+          ]}
+          onShareAddress={action('onShareAddress')}
+          onCopyAddress={action('onCopyAddress')}
+          isAddressValid={() => parseInt(Math.random() * 10, 10) > 3}
+          isIncentivizedTestnet={isIncentivizedTestnet}
+          currentLocale={currentLocale}
+          isShowingSubMenus
+        />
+        {showDialog && !isIncentivizedTestnet && (
+          <WalletReceiveDialog
+            address={generateAddress()}
+            onCopyAddress={action('onCopyAddress')}
+            onDownloadPDF={action('onDownloadPDF')}
+            onClose={action('onClose')}
+          />
         )}
-        isWalletAddressUsed={boolean('isWalletAddressUsed', false)}
-        walletAddresses={[
-          ...Array.from(Array(number('Addresses', 10))).map(() =>
-            generateAddress()
-          ),
-          ...Array.from(Array(number('Addresses (used)', 10))).map(() =>
-            generateAddress(true)
-          ),
-        ]}
-        onGenerateAddress={action('onGenerateAddress')}
-        onCopyAddress={action('onGenerateAddress')}
-        isSidebarExpanded={boolean('isSidebarExpanded', true)}
-        walletHasPassword={boolean('walletHasPassword', false)}
-        isSubmitting={boolean('isSubmitting', false)}
-      />
-    </VerticalFlexContainer>
-  ));
+      </VerticalFlexContainer>
+    );
+  });