+++ utils/v2-api-importer/index.js
@@ -2,11 +2,11 @@
 const axios = require('axios')
 
 const mnemonics = [
-  ['pass', 'proud', 'clarify', 'cargo', 'control', 'fancy', 'question', 'option', 'bring', 'recall', 'dolphin', 'meat', 'comic', 'version', 'pitch'],
-  ['behave', 'worth', 'later', 'banana', 'buzz', 'advance', 'local', 'owner', 'bulb', 'swear', 'theory', 'border', 'elephant', 'armor', 'chest'],
-  ['index', 'cover', 'candy', 'upper', 'orient', 'unit', 'shaft', 'peace', 'size', 'lava', 'wood', 'blush', 'huge', 'sunset', 'grid'],
-  ['liberty', 'powder', 'spoil', 'health', 'cable', 'problem', 'dice', 'alpha', 'solution', 'wagon', 'ask', 'siege', 'uncle', 'brick', 'life'],
-  ['shed', 'comic', 'subject', 'stage', 'dirt', 'yard', 'access', 'good', 'weasel', 'surround', 'since', 'climb', 'menu', 'blossom', 'joy'],
+  ['awkward', 'electric', 'strong', 'early', 'rose', 'abuse', 'mutual', 'limit', 'ketchup', 'child', 'limb', 'exist', 'hurry', 'business', 'whisper'],
+  ['blood', 'limit', 'pumpkin', 'fringe', 'order', 'trick', 'answer', 'festival', 'ethics', 'educate', 'luggage', 'dinner', 'record', 'fox', 'truth'],
+  ['bridge', 'joke', 'jeans', 'width', 'social', 'banner', 'visit', 'enlist', 'reason', 'hand', 'license', 'subway', 'butter', 'render', 'absent'],
+  ['bless', 'turkey', 'install', 'across', 'bronze', 'check', 'true', 'icon', 'treat', 'that', 'tuition', 'flush', 'panther', 'powder', 'ecology'],
+  ['trick', 'razor', 'bicycle', 'front', 'hollow', 'liberty', 'swift', 'coconut', 'pull', 'raccoon', 'level', 'woman', 'awful', 'sound', 'swarm'],
 ]
 
 const walletNames = [